Factor out common areas of service specs.
There was a lot of repetition in the service specs, so created a
base class that abstracts out a ton of it. Much cleaner now.
